added common library & split into classes
