+++ GildedRose.sln
@@ -12,6 +12,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		readme.md = readme.md
 	EndProjectSection
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GildedRoseKata.Common", "GildedRoseKata.Common\GildedRoseKata.Common.csproj", "{B2A4FC2F-C670-4433-B638-9B71B135B5BA}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -26,6 +28,10 @@ Global
 		{CB6715CE-A283-4C70-9C1B-F58822077731}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{CB6715CE-A283-4C70-9C1B-F58822077731}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{CB6715CE-A283-4C70-9C1B-F58822077731}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B2A4FC2F-C670-4433-B638-9B71B135B5BA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B2A4FC2F-C670-4433-B638-9B71B135B5BA}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B2A4FC2F-C670-4433-B638-9B71B135B5BA}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B2A4FC2F-C670-4433-B638-9B71B135B5BA}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
