+++ GildedRoseKata.Common/IQualityUpdater.cs
@@ -1,0 +1,8 @@
+﻿
+namespace GildedRoseKata.Common
+{
+    internal interface IQualityUpdater
+    {
+        void UpdateQuality();
+    }
+}
