+++ GildedRoseKata.Common/Item.cs
@@ -1,4 +1,4 @@
-﻿namespace GildedRoseKata
+﻿namespace GildedRoseKata.Common
 {
     public class Item
     {
